Modified attack formulas
This is a bit of a test for me figuring out pull requests. So here are a few changes I though about in the past days.

Using either STR or DEX/2 for the damage of light attack encouraged people to min-max and create DEX builds with 1 STR. Using the average of STR and DEX as an alternative to pure STR should still provide DEX builds an incentive to use Light, but now they'll profit from having a bit of STR too. Hopefully that will lead to more natural attribute spread. Reduced Stamina cost is there to compensate for light being the weakest attack and encourage people to use it more.

The rationale for the change to ranged is somewhat related. While the idea was to chose between weaker accurate attacks and stronger inaccurate attacks, that was not the case in practice. Everyone went either with very high DEX or if they had very high SPW they just went full mage. The proposed change would have ranged simulate two possibilities. Bows, guns, and thrown weapons all require strength IRL, so it makes sense to include it. The justification for higher STR giving more damage with guns is that it allows the fighter to use guns of a larger caliber. The DEX & SPW combo on the other hand would serve to represent attacks with energy weapons such as lasers or basic magical spells. The increased cost is there to make up for the potentially increased damage and because Ranged used to be too efficient in the amount of damage you can do with your stamina. This would put it halfway between Light and Heavy in cost, which seems fitting given that it's also halfway between them in damage.

The change to magic is to make the mage builds less extreme. Right now every other wizard has 10 SPW because it just scales the damage of Magic so much. A 10 SPW mage will do upwards of 30 damage even on poor rolls. Additionally a problem with Magic was that it used a single attribute for both attack and damage, which just added to the problem. The change to the attack table should bring it in line with all other attacks.

All this hurts mages considerably and I don't think they are actually too strong (win rate certainly doesn't suggest that), so making Magic cheaper should offset the reduced reliability and the new reliance on DEX for attacks.